Fixed rocket health stuff
The rocket health does not longer seem to desync, the problem was that each instance of rocket would cause damage to each instance of player, and depending on the latency/ping sometimes those were ignored/missed by the server overall. It's not all synced more properly with the server and only the instance of rocket actually shot by the player confirms with the server.